requirements.txt
fastapi
uvicorn
pydantic
fastapi[all]
openai
gtts
python-decouple
python-telegram-bot
